requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
awscli==1.38.37
backports.tarfile==1.2.0
boto3==1.37.37
botocore==1.37.37
bromo==0.2
captcha==0.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
config==0.5.1
contourpy==1.3.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cycler==0.12.1
diff-match-patch==20241021
dj-config-url==0.1.1
Django==5.1.5
django-autoslug==1.9.9
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-crispy-forms==2.3
django-csp==4.0
django-import-export==4.3.3
django-jalali==7.4.0
django-jalali-date==2.0.0
django-js-asset==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-storages==1.14.6
django-taggit==6.1.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
docutils==0.19
dotenv==0.9.9
ebcli==4.0.18
enum34==1.1.10
et_xmlfile==2.0.0
fields==5.0.0
fonttools==4.55.3
idna==3.10
import-export==0.3.1
jalali_core==1.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jdatetime==5.2.0
Jinja2==3.1.6
jmespath==1.0.1
keyring==9.0
keyrings.alt==5.0.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
postgres==4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.11.5
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
text-unidecode==1.3
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
